feat: new item edition page (images update works too)
