--- frontend/src/components/TheEditItemModal.vue
@@ -1,236 +1,0 @@
-<template>
-  <form @submit.prevent="save()">
-    <div class="modal-card">
-      <header class="modal-card-head">
-        <p class="modal-card-title">{{ $t('edit-item') }}</p>
-        <button class="delete" type="button" @click="$emit('close')" />
-      </header>
-      <section class="modal-card-body">
-        <b-field key="name" :message="errors.first('name')" :type="{'is-danger': errors.has('name')}">
-          <template #label>{{ $t('name') }}
-            <b-tooltip :label="$t('help_item_name')" multilined position="is-right">
-              <i class="icon far fa-question-circle"></i>
-            </b-tooltip>
-          </template>
-          <b-input v-model="internalItem.name" name="name" v-validate="'required'" />
-        </b-field>
-        <b-field key="type" :message="errors.first('type')" :type="{'is-danger': errors.has('type')}">
-          <template #label>
-            {{ $t('item-type') }}
-            <b-tooltip :label="$t('help_item_type')" multilined position="is-right">
-              <i class="icon far fa-question-circle"></i>
-            </b-tooltip>
-          </template>
-          <b-select v-model="internalItem.item_type" name="type" v-validate="'required'" expanded>
-            <option value="RQ">{{ $t('request') }}</option>
-            <option value="DN">{{ $t('donation') }}</option>
-            <option value="LN">{{ $t('loan') }}</option>
-            <option value="EV">{{ $t('event') }}</option>
-          </b-select>
-        </b-field>
-        <div class="columns">
-          <b-tooltip :label="$t('help_item_category')" multilined position="is-top">
-            <category-selector v-model="internalItem.category1" :number="1" class="column" expanded />
-            <category-selector v-model="internalItem.category2" :number="2" class="column" expanded />
-            <category-selector v-model="internalItem.category3" :number="3" class="column" expanded />
-          </b-tooltip>
-        </div>
-        <b-field>
-          <template #label>
-            {{ $t('address') }}
-            <b-tooltip :label="$t('help_item_address')" multilined position="is-right">
-              <i class="icon far fa-question-circle"></i>
-            </b-tooltip>
-          </template>
-          <b-input v-model="internalItem.address" />
-        </b-field>
-        <b-field key="description" :message="errors.first('description')" :type="{'is-danger': errors.has('description')}">
-          <template #label>
-            {{ $t('description') }}
-            <b-tooltip :label="$t('help_item_description')" multilined position="is-right">
-              <i class="icon far fa-question-circle"></i>
-            </b-tooltip>
-          </template>
-          <b-input v-model="internalItem.description" expanded type="textarea" name="description" v-validate="'required'" />
-        </b-field>
-        <b-field grouped>
-          <b-field key="startdate" :message="errors.first('startdate')" :type="{'is-danger': errors.has('startdate')}" expanded>
-            <template #label>
-              {{ $t('start-date') }}
-              <b-tooltip :label="$t('help_item_start_date')" multilined position="is-top">
-                <i class="icon far fa-question-circle"></i>
-              </b-tooltip>
-            </template>
-            <b-datetimepicker
-                v-model="internalItem.startdate"
-                :max-datetime="internalItem.enddate"
-                icon="calendar"
-                :icon-right="internalItem.startdate ? 'close-circle' : ''"
-                icon-right-clickable
-                @icon-right-click="clearStartdate"
-                icon-pack="fas"
-                :locale="$i18n.locale"
-                @change="startdateChanged"
-                name="startdate"
-                v-validate="'required'"
-            />
-          </b-field>
-          <b-field expanded>
-            <template #label> {{ $t('end-date') }}
-              <b-tooltip :label="$t('help_item_end_date')" multilined position="is-top">
-                <i class="icon far fa-question-circle"></i>
-              </b-tooltip>
-            </template>
-            <b-datetimepicker
-                v-model="internalItem.enddate"
-                :min-datetime="minEnddate"
-                icon="calendar"
-                :icon-right="internalItem.enddate ? 'close-circle' : ''"
-                icon-right-clickable
-                @icon-right-click="clearEnddate"
-                icon-pack="fas"
-                :locale="$i18n.locale"
-                @change="enddateChanged"
-            />
-          </b-field>
-        </b-field>
-        <b-field>
-          <b-checkbox v-model="internalItem.is_recurrent">
-            <strong>{{ $t('save-as-recurrent-item') }}</strong>
-            <b-tooltip :label="$t('help_item_recurrent')" multilined position="is-top">
-              <i class="icon far fa-question-circle"></i>
-            </b-tooltip>
-          </b-checkbox>
-        </b-field>
-      </section>
-      <footer class="modal-card-foot">
-        <b-button :label="$t('cancel')" @click="$emit('close')" />
-        <b-button :label="$t('save')" type="is-primary" :loading="waitingFormResponse" @click="save" />
-      </footer>
-    </div>
-  </form>
-</template>
-
-<script>
-import axios from "axios";
-import moment from "moment/moment";
-import CategorySelector from "@/components/CategorySelector";
-import ErrorHandler from "@/components/ErrorHandler";
-
-export default {
-  name: 'TheEditItemModal',
-  mixins: [ErrorHandler],
-  $_veeValidate: {
-    validator: 'new'
-  },
-  components: {CategorySelector},
-  props: {
-    item: {
-      type: Object,
-      required: true
-    },
-    address: String
-  },
-  data() {
-    return {
-      internalItem: {},
-      displayErrors: false,
-      waitingFormResponse: false,
-      initialStartdate: Date.now()
-    }
-  },
-  computed: {
-    minEnddate() {
-      const now = new Date();
-      if (this.internalItem.startdate) {
-        const startdate = new Date(this.internalItem.startdate);
-        return (startdate > now) ? startdate : now;
-      }
-      return now;
-    }
-  },
-  methods: {
-    async save() {
-      this.waitingFormResponse = true;
-
-      let result = await this.$validator.validateAll();
-      if (result) {
-        try {
-          let startdate;
-          if (this.internalItem.startdate)
-            startdate = moment(this.internalItem.startdate).format("YYYY-MM-DD[T]HH:mm:ss");
-          else
-            startdate = moment().format("YYYY-MM-DD[T]HH:mm:ss");
-
-          let enddate;
-          if (this.internalItem.enddate)
-            enddate = moment(this.internalItem.enddate).format("YYYY-MM-DD[T]HH:mm:ss");
-          else
-            enddate = null;
-
-          let item = (await axios.patch(`/api/v1/items/${this.item.id}/`, {
-            name: this.internalItem.name,
-            item_type: this.internalItem.item_type,
-            category1: this.internalItem.category1,
-            category2: this.internalItem.category2,
-            category3: this.internalItem.category3,
-            description: this.internalItem.description,
-            location: this.internalItem.address,
-            is_recurrent: this.internalItem.is_recurrent,
-            startdate: startdate,
-            enddate: enddate,
-          })).data;
-
-          this.$buefy.snackbar.open({
-            duration: 5000,
-            type: 'is-success',
-            message: this.$t('notif-success-item-update'),
-            pauseOnHover: true,
-          });
-          this.$emit('close');
-          this.$emit('updateItem', item);
-        }
-        catch (error) {
-          this.fullErrorHandling(error);
-        }
-      }
-
-      this.waitingFormResponse = false;
-    },
-    clearStartdate() {
-      this.internalItem.startdate = this.initialStartdate;
-      this.startdateChanged();
-    },
-    clearEnddate() {
-      this.internalItem.enddate = null;
-      this.enddateChanged();
-    },
-    startdateChanged() {
-      if (this.internalItem.enddate && this.internalItem.startdate) {
-        if (this.internalItem.startdate > this.internalItem.enddate)
-          this.internalItem.enddate = this.internalItem.startdate;
-      }
-    },
-    enddateChanged() {
-      if (this.internalItem.enddate && this.internalItem.startdate) {
-        if (this.internalItem.enddate < this.internalItem.startdate)
-          this.internalItem.startdate = this.internalItem.enddate;
-      }
-    }
-  },
-  mounted() {
-    this.internalItem = {...this.item};
-    
-    if (this.internalItem.startdate)
-      this.internalItem.startdate = new Date(this.internalItem.startdate);
-    this.initialStartdate = this.internalItem.startdate;
-
-    if (this.internalItem.enddate)
-      this.internalItem.enddate = new Date(this.internalItem.enddate);
-
-    this.internalItem.address = this.address;
-
-    delete this.internalItem.images;
-  }
-};
-</script>
