+++ frontend/src/locales/json/fr.i18n.json
@@ -145,7 +145,7 @@
   "language-fr": "Français",
   "to": "à",
   "too-small-zoom": "Zoomez pour voir plus de données",
-  "edit": "Editer",
+  "edit": "Éditer",
   "log-out": "Se déconnecter",
   "my-items": "Mes éléments",
   "firstname": "Prénom",
@@ -227,7 +227,7 @@
   "enter-new-password": "Entrez votre nouveau mot de passe ici...",
   "notif-success-password-reset": "Votre mot de passe a été modifié.",
   "notif-error-password-reset": "Une erreur a empêché la mise à jour de votre mot de passe.",
-  "edit-item": "Editer un élément",
+  "edit-item": "Éditer un élément",
   "notif-success-item-update": "L'élément a été correctement édité",
   "item-image": "Image",
   "item-image-info": "Uploadez une image pour remplacer l'actuelle.",
