+++ frontend/src/locales/translations.csv
@@ -146,7 +146,7 @@ language-en,English,Anglais
 language-fr,French,Français
 to,to,à
 too-small-zoom,Zoom to show more data,Zoomez pour voir plus de données
-edit,Edit,Editer
+edit,Edit,Éditer
 log-out,Log out,Se déconnecter
 my-items,My items,Mes éléments
 firstname,First name,Prénom
@@ -224,7 +224,7 @@ notif-error-password-reset-sent,An error prevented to reset your password.,Une e
 enter-new-password,Enter your new password here...,Entrez votre nouveau mot de passe ici...
 notif-success-password-reset,Your password has been changed.,Votre mot de passe a été modifié.
 notif-error-password-reset,An error prevented to change your password.,Une erreur a empêché la mise à jour de votre mot de passe.
-edit-item,Edit an item,Editer un élément
+edit-item,Edit an item,Éditer un élément
 notif-success-item-update,The item has been successfully edited,L'élément a été correctement édité
 item-image,Image,Image
 item-image-info,Only upload image to replace existing one.,Uploadez une image pour remplacer l'actuelle.
